cria condicoes e exception para tentativa de deletar estadio ja usado…

+++ src/main/java/com/example/federacao_develop/service/EstadioService.java
@@ -2,7 +2,9 @@ package com.example.federacao_develop.service;
 
 import com.example.federacao_develop.dto.EstadioDTO;
 import com.example.federacao_develop.model.Estadio;
+import com.example.federacao_develop.repository.ClubeRepository;
 import com.example.federacao_develop.repository.EstadioRepository;
+import com.example.federacao_develop.repository.PartidaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +16,11 @@ public class EstadioService {
 
     @Autowired
     private EstadioRepository estadioRepository;
+
+    @Autowired
+    private ClubeRepository clubeRepository;
+    @Autowired
+    private PartidaRepository partidaRepository;
 
     private EstadioDTO toDTO(Estadio estadio) {
         EstadioDTO dto = new EstadioDTO();
@@ -54,6 +61,14 @@ public class EstadioService {
     }
 
     public void delete(Integer id) {
+        boolean estadioTemClubes = clubeRepository.existsByEstadio_EstadioId(id);
+        boolean estadioTemPartidas = partidaRepository.existsByEstadio_EstadioId(id);
+        if (estadioTemClubes) {
+            throw new RuntimeException("Não é possível deletar o estádio pois existem clubes vinculados a ele. Para deletar este estádio, você precisa primeiro deletar os clubes associados a ele.");
+        }
+        if (estadioTemPartidas) {
+            throw new RuntimeException("Não é possível deletar o estádio pois existem partidas marcadas para ele. Para deletar este est[adio, você precisa primeiro deletar as partidas associadas a ele.");
+        }
         estadioRepository.deleteById(id);
     }
 }
